Extract component to formik-components folder
