+++ src/components/dashboard/daybook/DayBookTransactionForm/index.tsx
@@ -4,12 +4,10 @@ import {
   DialogActions,
   DialogContent,
   Stack,
-  TextField,
 } from '@mui/material';
-import { DatePicker, DatePickerProps } from '@mui/x-date-pickers';
-import { Form, Formik, FormikConfig, useFormikContext } from 'formik';
-import { get } from 'lodash';
+import { Form, Formik, FormikConfig } from 'formik';
 import { FC } from 'react';
+import { FormikDatePicker } from 'src/components/shared/formik-components';
 import { DayBookTransaction } from 'src/types/dayBook';
 import { DayBookTransactionsTable } from './DayBookTransactionsTable';
 
@@ -57,35 +55,5 @@ export const DayBookTransactionForm: FC<DayBookTransactionFormProps> = (
         </Form>
       )}
     </Formik>
-  );
-};
-
-interface FormikDatePickerProps
-  extends Omit<
-    DatePickerProps<null, null>,
-    'value' | 'onChange' | 'renderInput'
-  > {
-  name: string;
-}
-
-const FormikDatePicker: FC<FormikDatePickerProps> = (props) => {
-  const { name } = props;
-  const { values, touched, errors, setFieldValue } = useFormikContext();
-
-  const fieldValue = get(values, name, '');
-  const fieldTouched = get(touched, name, false);
-  const fieldError = get(errors, name, '');
-
-  const handleChange = (value: string | null) => {
-    setFieldValue(name, value);
-  };
-  return (
-    <DatePicker
-      value={fieldValue}
-      onChange={handleChange}
-      renderInput={(params) => (
-        <TextField {...params} error={fieldTouched} helperText={fieldError} />
-      )}
-    />
   );
 };
