+++ src/components/shared/formik-components/index.ts
@@ -1,2 +1,3 @@
 export { FormikAutoCalculateField } from './FormikAutoCalculateField';
+export { FormikDatePicker } from './FormikDatePicker';
 export { FormikTextField } from './FormikTextField';
